Refactored SmartiRelaod
