+++ packages/assistify-ai/server/lib/SmartiAdapter.js
@@ -76,31 +76,16 @@ export class SmartiAdapter {
 			}
 			if (request_result) {
 				SystemLogger.debug('Conversation found and message will be synced now');
-				Meteor.defer(()=>Meteor.call('markMessageAsSynced', message._id));
+				Meteor.defer(() => SmartiAdapter._markMessageAsSynced(message._id));
 			} else {
-				Meteor.defer(()=>Meteor.call('markRoomAsUnsynced', message.rid));
+				Meteor.defer(() => SmartiAdapter._markRoomAsUnsynced(message.rid));
 			}
 		} else {
 			// create conversation
 			SystemLogger.debug(`Conversation not found for room ${ message.rid }, create a new conversation.`);
-			const helpRequest = RocketChat.models.HelpRequests.findOneByRoomId(message.rid);
+
 			const room = RocketChat.models.Rooms.findOneById(message.rid);
-
-			// The "support_area" in Smarti is an optional property. A historic conversation belonging to the same support_are increases relevance
-			let supportArea = room.parentRoomId || room.topic || room.expertise;
-			if (!supportArea) {
-				if (room.t === '') {
-					supportArea = 'livechat';
-				} else if (helpRequest && helpRequest.supportArea) {
-					supportArea = helpRequest.supportArea;
-				} else {
-					supportArea = room.name;
-				}
-			}
-
-			SystemLogger.debug('HelpRequest:', helpRequest);
-			SystemLogger.debug('Room:', room);
-
+			const supportArea = this._getSupportArea(room);
 			const requestBodyConversation = {
 				'meta': {
 					'support_area': [supportArea],
@@ -124,9 +109,9 @@ export class SmartiAdapter {
 			if (conversation && conversation.id) {
 				conversationId = conversation.id;
 				SystemLogger.debug('New conversation created and message will be synced now');
-				Meteor.defer(()=>Meteor.call('markMessageAsSynced', message._id));
+				Meteor.defer(() => SmartiAdapter._markMessageAsSynced(message._id));
 			} else {
-				Meteor.defer(()=>Meteor.call('markRoomAsUnsynced', message.rid));
+				Meteor.defer(() => SmartiAdapter._markRoomAsUnsynced(message.rid));
 			}
 		}
 
@@ -147,8 +132,7 @@ export class SmartiAdapter {
 
 		const conversationId = SmartiAdapter.getConversationId(message.rid);
 		if (conversationId) {
-			SystemLogger.debug(`Conversation ${ conversationId } found for channel ${ message.rid }`);
-			SystemLogger.debug(`Deleting message from conversation ${ conversationId } ...`);
+			SystemLogger.debug(`Smarti - Deleting message ${ message.rid } from conversation ${ conversationId }.`);
 			SmartiProxy.propagateToSmarti(verbs.delete, `conversation/${ conversationId }/message/${ message._id }`);
 			SmartiAdapter._getAnalysisResult(message.rid, conversationId);
 		} else {
@@ -165,6 +149,7 @@ export class SmartiAdapter {
 
 		const conversationId = SmartiAdapter.getConversationId(room._id);
 		if (conversationId) {
+			SystemLogger.debug(`Smarti - Deleting conversation ${ conversationId } after room ${ room._id  } erased.`);
 			SmartiProxy.propagateToSmarti(verbs.delete, `/conversation/${ conversationId }`);
 			SmartiAdapter._removeMapping(room.id);
 		} else {
@@ -186,35 +171,11 @@ export class SmartiAdapter {
 		if (conversationId) {
 			const res = SmartiProxy.propagateToSmarti(verbs.put, `/conversation/${ conversationId }/meta.status`, 'Complete');
 			if (!res) {
-				Meteor.defer(()=>Meteor.call('markRoomAsUnsynced', room._id));
+				Meteor.defer(() => SmartiAdapter._markRoomAsUnsynced(room._id));
 			}
 		} else {
 			SystemLogger.error(`Smarti - setting conversation to complete faild when closing room [ ${ room._id } ]`);
 		}
-	}
-
-	/**
-	 * At any point in time when the anylsis for a room has been done (onClose, onMessage), this analysisCompleted
-	 * Updates the mapping and notifies the room, in order to reload the Smarti result representation
-	 *
-	 * @param roomId - the RC room Id
-	 * @param conversationId  - the Smarti conversation Id
-	 * @param analysisResult - the analysis result from Smarti
-	 */
-	static analysisCompleted(roomId, conversationId, analysisResult) {
-
-		if (roomId === null) {
-			const conversationCacheEntry = RocketChat.models.LivechatExternalMessage.findOneByConversationId(conversationId);
-			if (conversationCacheEntry && conversationCacheEntry.rid) {
-				roomId = conversationCacheEntry.rid;
-			} else {
-				SystemLogger.error(`Smarti - no room found for conversation [ ${ conversationId } ] unable to notify room.`);
-			}
-		}
-		SystemLogger.debug('Smarti - retieved analysis result =', JSON.stringify(analysisResult, null, 2));
-		SystemLogger.debug(`Smarti - analysis complete -> update cache and notify room [ ${ roomId } ] for conversation [ ${ conversationId } ]`);
-		SmartiAdapter._updateMapping(roomId, conversationId);
-		RocketChat.Notifications.notifyRoom(roomId, 'newConversationResult', analysisResult);
 	}
 
 	/**
@@ -256,8 +217,44 @@ export class SmartiAdapter {
 		return conversationId;
 	}
 
-	static resynch(ignoreSynchFlag) {
+	/**
+	 * At any point in time when the anylsis for a room has been done (onClose, onMessage), this analysisCompleted
+	 * Updates the mapping and notifies the room, in order to reload the Smarti result representation
+	 *
+	 * @param roomId - the RC room Id
+	 * @param conversationId  - the Smarti conversation Id
+	 * @param analysisResult - the analysis result from Smarti
+	 */
+	static analysisCompleted(roomId, conversationId, analysisResult) {
+
+		if (roomId === null) {
+			const conversationCacheEntry = RocketChat.models.LivechatExternalMessage.findOneByConversationId(conversationId);
+			if (conversationCacheEntry && conversationCacheEntry.rid) {
+				roomId = conversationCacheEntry.rid;
+			} else {
+				SystemLogger.error(`Smarti - no room found for conversation [ ${ conversationId } ] unable to notify room.`);
+			}
+		}
+		SystemLogger.debug('Smarti - retieved analysis result =', JSON.stringify(analysisResult, null, 2));
+		SystemLogger.debug(`Smarti - analysis complete -> update cache and notify room [ ${ roomId } ] for conversation [ ${ conversationId } ]`);
+		SmartiAdapter._updateMapping(roomId, conversationId);
+		RocketChat.Notifications.notifyRoom(roomId, 'newConversationResult', analysisResult);
+	}
+
+	static _getAnalysisResult(roomId, conversationId) {
+
+		// conversation updated or created => request analysis results
+		SmartiAdapter._updateMapping(roomId, conversationId);
+		SystemLogger.debug(`Smarti - conversation updated or created -> get analysis result asynch [ callback=${ SmartiAdapter.rocketWebhookUrl } ] for conversation: ${ conversationId } and room: ${ roomId }`);
+		SmartiProxy.propagateToSmarti(verbs.get, `conversation/${ conversationId }/analysis?callback=${ SmartiAdapter.rocketWebhookUrl }`); // asynch
+	}
+
+	static resync(ignoreSynchFlag) {
 		SystemLogger.info('Smarti resync triggered');
+
+		if (ignoreSynchFlag) {
+			RocketChat.models.LivechatExternalMessage.clear();
+		}
 
 		let query = {};
 		if (!ignoreSynchFlag || ignoreSynchFlag !== true) {
@@ -268,14 +265,14 @@ export class SmartiAdapter {
 		const requests = RocketChat.models.Rooms.model.find(query).fetch();
 		SystemLogger.info('Number of Requests to sync: ', requests.length);
 		for (let i = 0; i < requests.length; i++) {
-			Meteor.defer(() => Meteor.call('tryResync', requests[i]._id, ignoreSynchFlag));
+			Meteor.defer(() => SmartiAdapter._tryResync(requests[i]._id, ignoreSynchFlag));
 		}
 
 		query.t = 'e';
 		const topics = RocketChat.models.Rooms.model.find(query).fetch();
 		SystemLogger.info('Number of Topics to sync: ', topics.length);
 		for (let i = 0; i < topics.length; i++) {
-			Meteor.defer(() => Meteor.call('tryResync', topics[i]._id), ignoreSynchFlag);
+			Meteor.defer(() => SmartiAdapter._tryResync(topics[i]._id, ignoreSynchFlag));
 		}
 
 		return {
@@ -283,12 +280,141 @@ export class SmartiAdapter {
 		};
 	}
 
-	static _getAnalysisResult(roomId, conversationId) {
+	static _tryResync(rid, ignoreSynchFlag) {
+		SystemLogger.debug('Sync all unsynced messages in room: ', rid);
+		const room = RocketChat.models.Rooms.findOneById(rid);
+		const messageDB = RocketChat.models.Messages;
+		const unsync = messageDB.find({ lastSync: { $exists: false }, rid, t: { $exists: false } }).fetch();
+		if (!ignoreSynchFlag || ignoreSynchFlag !== true) {
+			if (unsync.length === 0 && !room.outOfSync) {
+				SystemLogger.debug('Room is already in sync');
+				return true;
+			} else {
+				SystemLogger.debug('Messages out of sync: ', unsync.length);
+			}
+		}
 
-		// conversation updated or created => request analysis results
-		SmartiAdapter._updateMapping(roomId, conversationId);
-		SystemLogger.debug(`Smarti - conversation updated or created -> get analysis result asynch [ callback=${ SmartiAdapter.rocketWebhookUrl } ] for conversation: ${ conversationId } and room: ${ roomId }`);
-		SmartiProxy.propagateToSmarti(verbs.get, `conversation/${ conversationId }/analysis?callback=${ SmartiAdapter.rocketWebhookUrl }`); // asynch
+		SmartiProxy.propagateToSmarti(verbs.get, 'system/info', null, (response) => {
+			// if Smarti is not available stop
+			if (response.statusCode !== 200) {
+				throw new Meteor.Error('Smarti not reachable');
+			}
+		});
+
+		const conversationId = SmartiAdapter.getConversationId(rid);
+		if (conversationId) {
+			SystemLogger.debug(`Conversation found ${ conversationId } - delete and create new conversation`);
+			SmartiProxy.propagateToSmarti(verbs.delete, `conversation/${ conversationId }`, null);
+		}
+
+		// create the conversation completly new (this will also heal defected conversations)
+		const supportArea = SmartiAdapter._getSupportArea(room);
+		const conversation = {
+			'meta': {
+				'support_area': [supportArea],
+				'channel_id': [rid]
+			},
+			'user': {
+				'id': room.u._id
+			},
+			'messages': [],
+			'context': {
+				'contextType': 'rocket.chat'
+			}
+		};
+		let messages;
+		if (room.closedAt) {
+			messages = messageDB.find({rid, ts: {$lt: room.closedAt}}, { sort: { ts: 1 } }).fetch();
+			conversation.meta.status = 'Complete';
+		} else {
+			messages = messageDB.find({rid}, { sort: { ts: 1 } }).fetch();
+		}
+
+		conversation.messages = [];
+		for (let i=0; i < messages.length; i++) {
+			const newMessage = {
+				'id': messages[i]._id,
+				'time': messages[i].ts,
+				'origin': 'User', //user.type,
+				'content': messages[i].msg,
+				'user': {
+					'id': messages[i].u._id
+				}
+			};
+			conversation.messages.push(newMessage);
+		}
+
+		// post the new conversation to Smarti
+		const newSmartiConversation = SmartiProxy.propagateToSmarti(verbs.post, 'conversation', conversation);
+		// get the analysis result
+		const analysisResult = SmartiProxy.propagateToSmarti(verbs.get, `conversation/${ newSmartiConversation.id }/analysis`);
+		SystemLogger.debug(`Smarti - New conversation with Id ${ newSmartiConversation.id } analyzed.`);
+		if (analysisResult) {
+			SystemLogger.debug('analysisResult:', JSON.stringify(analysisResult, null, '\t'));
+			SmartiAdapter.analysisCompleted(rid, newSmartiConversation.id, analysisResult);
+		}
+		SystemLogger.debug(`Smarti analysisCompleted finished for conversation ${ newSmartiConversation.id }`);
+
+		for (let i=0; i < messages.length; i++) {
+			Meteor.defer(() => SmartiAdapter._markMessageAsSynced(messages[i]._id));
+		}
+		RocketChat.models.Rooms.model.update(
+			{_id: rid},
+			{
+				$set: {
+					outOfSync: false
+				}
+			});
+		SystemLogger.debug('Room Id: ', rid, ' is in sync now');
+		return true;
+	}
+
+	static _getSupportArea(room) {
+		const helpRequest = RocketChat.models.HelpRequests.findOneByRoomId(room._id);
+		// The "support_area" in Smarti is an optional property. A historic conversation belonging to the same support_are increases relevance
+		let supportArea = room.parentRoomId || room.topic || room.expertise;
+		if (!supportArea) {
+			if (room.t === '') {
+				supportArea = 'livechat';
+			} else if (helpRequest && helpRequest.supportArea) {
+				supportArea = helpRequest.supportArea;
+			} else {
+				supportArea = room.name;
+			}
+		}
+
+		SystemLogger.debug('HelpRequest:', helpRequest);
+		SystemLogger.debug('Room:', room);
+		return supportArea;
+	}
+
+	static _markMessageAsSynced(messageId) {
+		const messageDB = RocketChat.models.Messages;
+		const message = messageDB.findOneById(messageId);
+		const lastUpdate = message ? message._updatedAt : 0;
+		if (lastUpdate) {
+			messageDB.model.update(
+				{_id: messageId},
+				{
+					$set: {
+						lastSync: lastUpdate
+					}
+				});
+			SystemLogger.debug('Message Id: ', messageId, ' has been synced');
+		} else {
+			SystemLogger.debug('Message Id: ', messageId, ' can not be synced');
+		}
+	}
+
+	static _markRoomAsUnsynced(rid) {
+		RocketChat.models.Rooms.model.update(
+			{_id: rid},
+			{
+				$set: {
+					outOfSync: true
+				}
+			});
+		SystemLogger.debug('Room Id: ', rid, ' is out of sync');
 	}
 
 	static _removeMapping(roomId) {
